♻️ refactor(monitoring): extract codex inspection presentation
Move Codex inspection page formatting helpers and panel components into dedicated presentation modules.

This reduces the page component without touching execution behavior; validated with type-check, codex inspection tests, and lint.
